Feat: added support for the %n specifier.
A big change is in the fact that now the current amount of written chars is passed through the
'handle_format' and the 'ouptut' functions for them to be able to pass
it to 'f_n'. When a buffer is implemented this should not have to be the
case anymore because then the buffer will keep track of 'n'.